Use SSLContext.getDefault() more now that it's guaranteed fixed
Without this, some of the existing advice logic needs to be able to reach
HttpProxyAgent, which fails in some environments. This avoids that entirely,
making runtime logic everywhere simpler & more reliable.

Also starts removing unnecessary extra try/catches, as it appears that
bytebuddy handles this for us by default anyway.