Make Apache v4 interception reliable enough to intercept Gradle
Not really because Gradle is especially interesting, more that it's sufficiently
common that it's likely to be a test case for many people, and it would be good
if that worked! Previously it did basically work, AFAICT, but with loud errors
during setup when HttpRoute couldn't be resolved. We now ensure that that's not
required, so there's no errors: just perfectly intercepted HTTP(S).