Make Scala interception more reliable for existing connection pools
Previously the tests would frequently fail because the pool wasn't
properly reset when interception began. This seems to be a race
condition if there's a request currently in progress, and it's something
that would likely resolve in practice _eventually_ but it's better to
detect cases and fix it properly per pool, rather than relying on
shutdownAll working reliably with concurrent requests.

Still more work to do here to make Akka reliable with older versions,
and check a couple of other edge cases.