Ensure Akka HTTP(S) interception correctly handles plain HTTP
If test URL is changed to plain HTTP the tests fail, but that's
expected, as Wiremock only supports HTTPS and Akka & Spring don't
support HTTP-over-HTTPS proxying correctly. HTTP Toolkit supports
both on the same port just fine though, so that's enough for now.

Might be worth investigating this further in future, but it's quite
tricky, not worth it until it causes a real world issue.