Fix minification of ByteBuddy 1.12.6 by disabling optimizations
This allows us to limit the extra entry points to just the agent code.
It does increase the JAR size by about 0.1MB (now 2.3MB in total) but it
should be significantly less fragile and it's still much smaller than
the raw JAR (4.8MB). 30% size increase, but at least half due to a real
increase in size of ByteBuddy itself.